+++ AlbumArt3.0@claudiux/files/AlbumArt3.0@claudiux/get-image-size.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+FILE=$1
+sleep 1
+RES=$(file ${FILE:7} | grep -Eo ", [[:digit:]]+ *x *[[:digit:]]+" | tr "," " ")
+
+echo -n $RES
+
+exit 0