+++ AlbumArt3.0@claudiux/files/AlbumArt3.0@claudiux/CHANGELOG.md
@@ -1,3 +1,12 @@
+### v2.1.0~20250314
+  * Added option "Activate Album Art display at startup".
+
+### v2.0.0~20250313
+  * Also works with the Enhanced Sound applet.
+  * Many changes.
+  * Works entirely in memory (RAM) to preserve SSD.
+  * Tries to respect the image's aspect ratio as far as possible.
+
 ### v1.0.0~20250211
   * Desklet used by the Radio3.0@claudiux applet to display the album art of the song being listened to, if available.
   * Creation.